requirements.txt
jiwer
pandas
numpy
tensorflow-gpu
string
glob
librosa
IPython